requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf
google-cloud-aiplatform>=1.38


streamlit
PyPDF2
pandas
seaborn
numpy
chardet
matplotlib
reportlab


pandas
numpy
matplotlib
seaborn
scipy
os-sys
statsmodels
PyPDF2
fpdf
sklearn